+++ deployment/ccip/shared/stateview/state.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
-	"strings"
 	"sync"
 
 	"github.com/Masterminds/semver/v3"
@@ -26,8 +25,6 @@ import (
 	cldf_evm "github.com/smartcontractkit/chainlink-deployments-framework/chain/evm"
 	cldf_chain_utils "github.com/smartcontractkit/chainlink-deployments-framework/chain/utils"
 	cldf "github.com/smartcontractkit/chainlink-deployments-framework/deployment"
-
-	evmstate "github.com/smartcontractkit/chainlink/deployment/common/changeset/state"
 
 	"github.com/smartcontractkit/chainlink-evm/gethwrappers/generated/link_token_interface"
 	"github.com/smartcontractkit/chainlink-evm/gethwrappers/shared/generated/initial/link_token"
@@ -794,7 +791,7 @@ func LoadOnchainState(e cldf.Environment) (CCIPOnChainState, error) {
 		evmMu:       &sync.RWMutex{},
 	}
 	for chainSelector, chain := range e.BlockChains.EVMChains() {
-		addresses, err := mergeAddressesFromBothSources(e, chainSelector)
+		addresses, err := commonstate.MergeAddressesFromBothSources(e, chainSelector)
 		if err != nil {
 			return state, fmt.Errorf("failed to merge addresses for chain %d: %w", chainSelector, err)
 		}
@@ -1411,49 +1408,4 @@ func LoadOnchainStateSolana(e cldf.Environment) (CCIPOnChainState, error) {
 		state.SolChains[chainSelector] = chainState
 	}
 	return state, nil
-}
-
-// mergeAddressesFromBothSources combines addresses from both DataStore and AddressBook making it backward compatible.
-// This should be deleted in future when AddressBook is fully deprecated
-// DataStore addresses take precedence over AddressBook addresses when there are conflicts
-func mergeAddressesFromBothSources(e cldf.Environment, chainSelector uint64) (map[string]cldf.TypeAndVersion, error) {
-	// Start with addresses from AddressBook for backward compatibility
-	addressBookAddresses := make(map[string]cldf.TypeAndVersion)
-	if addresses, err := e.ExistingAddresses.AddressesForChain(chainSelector); err == nil {
-		addressBookAddresses = addresses
-	} else if !errors.Is(err, cldf.ErrChainNotFound) {
-		return nil, fmt.Errorf("failed to load addresses from AddressBook: %w", err)
-	}
-
-	// Try to load addresses from DataStore (without qualifier for general case)
-	// Only try if DataStore is available
-	if e.DataStore != nil {
-		dataStoreAddresses, err := evmstate.LoadAddressesFromDataStore(e.DataStore, chainSelector, "")
-		if err != nil {
-			// If DataStore has no addresses, just return AddressBook addresses
-			if strings.Contains(err.Error(), "no addresses found") {
-				return addressBookAddresses, nil
-			}
-			// Don't fail if DataStore is not working - fall back to AddressBook
-			return addressBookAddresses, nil
-		}
-
-		// Merge the two maps - DataStore addresses take precedence
-		mergedAddresses := make(map[string]cldf.TypeAndVersion)
-
-		// First add all AddressBook addresses
-		for addr, tv := range addressBookAddresses {
-			mergedAddresses[addr] = tv
-		}
-
-		// Then add DataStore addresses (overwriting any conflicts)
-		for addr, tv := range dataStoreAddresses {
-			mergedAddresses[addr] = tv
-		}
-
-		return mergedAddresses, nil
-	}
-
-	// If no DataStore, just return AddressBook addresses
-	return addressBookAddresses, nil
 }
